delete waarschuwingen

+++ src/app/modules/places/places/places.component.ts
@@ -23,49 +23,49 @@ export class PlacesComponent implements OnInit {
   tables: TableLocation[];
   places: Place[];
 
-  displayedColumns = ["table.location.name","table.name","name", 'btn'];
+  displayedColumns = ["table.location.name", "table.name", "name", 'btn'];
   dataSource = new MatTableDataSource<Place>();
   @ViewChild(MatPaginator) paginator: MatPaginator;
 
-  constructor(private _tablesService : TablesService,private _placesService: PlacesService,private _locationsService:LocationsService, private router: Router) {
+  constructor(private _tablesService: TablesService, private _placesService: PlacesService, private _locationsService: LocationsService, private router: Router) {
     this._placesService.getPlaces().subscribe(
       result => {
         this.dataSource = new MatTableDataSource<Place>(result);
         this.dataSource.sortingDataAccessor = (item, property) => {
           switch (property) {
-              case 'table.location.name': return item.tableLocation.location.name;
-              case 'table.name': return item.tableLocation.name;
-              default: return item[property];
+            case 'table.location.name': return item.tableLocation.location.name;
+            case 'table.name': return item.tableLocation.name;
+            default: return item[property];
           }
         }
         this.dataSource.filterPredicate = (data: Place, filter: string) => {
           return data.name.toLocaleLowerCase().includes(filter) ||
-          data.tableLocation.name.toLocaleLowerCase().includes(filter) ||
-          data.tableLocation.location.name.toLocaleLowerCase().includes(filter);
+            data.tableLocation.name.toLocaleLowerCase().includes(filter) ||
+            data.tableLocation.location.name.toLocaleLowerCase().includes(filter);
         }
         this.dataSource.paginator = this.paginator;
         this.dataSource.sort = this.sort;
-        console.log(this.dataSource);  
+        console.log(this.dataSource);
       }
     )
     this._locationsService.getLocations().subscribe(
       result => {
-          this.locations = result
+        this.locations = result
       }
     )
     this._placesService.getPlaces().subscribe(
       result => {
-          this.places = result
+        this.places = result
       }
     )
-   }
-
-   @ViewChild(MatSort) sort: MatSort;
-   ngAfterViewInit() {
-     this.dataSource.paginator = this.paginator;
-     this.dataSource.sort = this.sort;
   }
-   applyFilter(filterValue: string) {
+
+  @ViewChild(MatSort) sort: MatSort;
+  ngAfterViewInit() {
+    this.dataSource.paginator = this.paginator;
+    this.dataSource.sort = this.sort;
+  }
+  applyFilter(filterValue: string) {
     filterValue = filterValue.trim(); // Remove whitespace
     filterValue = filterValue.toLowerCase(); // Datasource defaults to lowercase matches
     this.dataSource.filter = filterValue;
@@ -75,21 +75,41 @@ export class PlacesComponent implements OnInit {
   ngOnInit(): void {
     this._tablesService.getTables().subscribe(
       result => {
-      this.tables = result;
-    })
+        this.tables = result;
+      })
   }
 
   addPlace() {
-    let place = new Place(0,"EmptyPlace",this.tables[0],false);
+    let place = new Place(0, "EmptyPlace", this.tables[0], false);
     this._placesService.setEditPlace(place)
     //console.log(this._placesService.getPlace().id + ": "+ this._placesService.getPlace().name+","+this._placesService.getPlace().tableLocation.name)
     this.router.navigate(['/places/edit']);
   }
 
-  deletePlace(id: number) {
-    this._placesService.deletePlace(id).subscribe();
-    window.location.reload();
-    this.router.navigate(['/places']);
+  deletePlace(id: number, name: string) {
+    if (window.confirm("Wil je deze plaats: " + name +  " verwijderen?")) {
+      this._placesService.deletePlace(id).subscribe();
+      setTimeout(() => {
+        this._placesService.getPlaces().subscribe(
+          result => {
+            this.dataSource = new MatTableDataSource<Place>(result);
+            this.dataSource.sortingDataAccessor = (item, property) => {
+              switch (property) {
+                case 'table.location.name': return item.tableLocation.location.name;
+                case 'table.name': return item.tableLocation.name;
+                default: return item[property];
+              }
+            }
+            this.dataSource.filterPredicate = (data: Place, filter: string) => {
+              return data.name.toLocaleLowerCase().includes(filter) ||
+                data.tableLocation.name.toLocaleLowerCase().includes(filter) ||
+                data.tableLocation.location.name.toLocaleLowerCase().includes(filter);
+            }
+            this.dataSource.paginator = this.paginator;
+            this.dataSource.sort = this.sort;
+          })
+      }, 1000);
+    }
   }
 
   showDetailPlace(place: Place) {
